🚚Rename: Modal 관련 파일 이동 및 import문 수정
