--- src/ui/common/TripModal.tsx
@@ -1,3 +1,0 @@
-export default function TripModal() {
-  return <section>여행 생성/수정 모달</section>;
-}
